requirements.txt
Django~=5.2.3
django-tables2~=2.7.5
django-guardian~=3.0.3